% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_area.R
\name{poly_area}
\alias{poly_area}
\title{Calculate polygon area by category within buffers of pointss}
\usage{
poly_area(points, IDs, bufferSize, polygons, categories,
  standardize = FALSE)
}
\arguments{
\item{points}{Spatial points object}

\item{IDs}{Unique ID field in points}

\item{bufferSize}{Radius used to buffer points.}

\item{polygons}{Spatial polygons object, which contains land use classes}

\item{categories}{Land use classes in the spatial polygon object}

\item{standardize}{Row Standardize area calculations to proportions instead of area}
}
\description{
For a set of points, buffer each point and calculate the amount of area by land use category
}
