% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_length.R
\name{line_length}
\alias{line_length}
\title{Calculate line length by category within buffers of points}
\usage{
line_length(points, IDs, bufferSize, lines, categories)
}
\arguments{
\item{points}{Spatial points object}

\item{IDs}{Unique ID field in points}

\item{bufferSize}{Radius used to buffer points.}

\item{lines}{Spatial lines object, which contains road classes}

\item{categories}{Road types in the spatial line object}
}
\description{
For a set of points, buffer each point and calculate the total length of
surronding lines by type. Often used to calculate the length of road by type.
}
\details{
The length of road by type is a common variable in land use regression models.
}
